feat: init struture for moeda, controller, repository, service, dtos,…

+++ back-poly/polyrate/src/main/java/nick/project/polyrate/Moeda/Entity/MoedaEntity.java
@@ -15,7 +15,7 @@ public class MoedaEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_moeda")
-    private BigInteger idMoeda;
+    private Integer idMoeda;
 
     @Column
     private String codigo;
